DRTII-1165 Plug in Lgw forecast using sftp
