+++ server/src/main/resources/config/feeds.conf
@@ -44,9 +44,15 @@ feeds {
       }
     }
     forecast {
-      userId = ${?LGW_BOX_USER_ID}
-      folderId = ${?LGW_BOX_FOLDER_ID}
-      boxConfigFile = ${?LGW_BOX_CONFIG_FILE}
+      sftp {
+        username = ${?LGW_SFTP_USERNAME}
+        password = ${?LGW_SFTP_PASSWORD}
+        host = ${?LGW_SFTP_HOST}
+        pathPrefix = 'inbox'
+        pathPrefix = ${?LGW_SFTP_PATH_PREFIX}
+      }
+      initial-delay-seconds = 10
+      interval-minutes = 1440
     }
   }
