+++ server/src/main/scala/drt/server/feeds/lgw/LgwForecastSftpService.scala
@@ -1,0 +1,75 @@
+package drt.server.feeds.lgw
+
+import drt.server.feeds.lgw.LgwForecastSftpService.sshClient
+import net.schmizz.sshj.SSHClient
+import net.schmizz.sshj.sftp.SFTPClient
+import net.schmizz.sshj.transport.verification.PromiscuousVerifier
+import net.schmizz.sshj.xfer.InMemoryDestFile
+import org.slf4j.{Logger, LoggerFactory}
+import uk.gov.homeoffice.drt.time.SDate
+
+import java.io.ByteArrayOutputStream
+import java.nio.charset.StandardCharsets
+import scala.jdk.CollectionConverters.CollectionHasAsScala
+import scala.util.{Failure, Success, Try}
+
+
+object LgwForecastSftpService {
+  def sshClient(ftpServer: String, username: String, password: String): SSHClient = {
+    val ssh = new SSHClient()
+    ssh.loadKnownHosts()
+    ssh.addHostKeyVerifier(new PromiscuousVerifier())
+    ssh.connect(ftpServer)
+    ssh.authPassword(username, password)
+    ssh.setTimeout(0)
+    ssh
+  }
+}
+
+case class LgwForecastSftpService(ftpServer: String, username: String, password: String, pathPrefix: String) {
+  private val log: Logger = LoggerFactory.getLogger(getClass)
+
+  def ssh: SSHClient = sshClient(ftpServer, username, password)
+
+  val latestContent: () => Option[String] = () => {
+    val client = ssh.newSFTPClient()
+    val maybeContent = latestFileName(client).flatMap(file => contentForFile(client, file))
+    client.close()
+    maybeContent
+  }
+
+  private def contentForFile(sftp: SFTPClient, fileName: String): Option[String] = {
+    val outputStream: ByteArrayOutputStream = new ByteArrayOutputStream()
+
+    val output: InMemoryDestFile = new InMemoryDestFile {
+      def getOutputStream: ByteArrayOutputStream = outputStream
+    }
+
+    Try(sftp.get(fileName, output)) match {
+      case Success(_) => Some(outputStream.toString(StandardCharsets.UTF_8))
+      case Failure(e) =>
+        log.error(s"Failed to get file '$fileName' from sftp server: ${e.getMessage}")
+        None
+    }
+  }
+
+  private def latestFileName(sftp: SFTPClient): Option[String] = {
+    sftp
+      .ls(pathPrefix).asScala
+      .filter(_.getName.contains("-LGWArrForecast.csv"))
+      .map { f =>
+        Try(f.getName.split("-", 4)).map {
+          case Array(year, month, day, _) =>
+            (SDate(year.toInt, month.toInt, day.toInt), f.getPath)
+        }
+      }
+      .collect {
+        case Success(someDate) => someDate
+      }
+      .toList
+      .sortBy(_._1.millisSinceEpoch)
+      .reverse
+      .headOption
+      .map(_._2)
+  }
+}
